Prevent error when handling a non-Blazar exception
The code in this exception handler access the restore_lease_status
attribute of BlazarException objects. However, since we are handling any
Exception object, this attribute could be missing if the exception
doesn't inherit from BlazarException, for example for the Unauthorized
exception.

Use getattr() to check that there actually is a restore_lease_status
attribute.