Add on_start mechanism and use to launch stack
This adds a new 'on_start' behavior that can be customized on a lease.
Currently there is only one value, 'orchestration:<stack_id>', which
will trigger the launch of a Heat stack with the given ID when the lease
starts. The reservation_id will be passed to the Heat stack as a
parameter.