Add date lease info to host allocations list
Allocations API can be used by admins to see usage and availablility of
resources. However, due to RBAC for lease list functionality, it is not
possibility users to use allocations API for this purpose. This adds
start_date and end_date for users so that they can use allocations api
to see usage and availability of resources.