Bug Fix: Error Host + Network Lease
If a lease contains both network and host reservations while the
host has a live instance on the network, the lease will fail on
termination. This can occur either because the network tear down
occurs first or because the vif on the ironic port persists for
some time shortly after instance is deleted.

This patch ensures networks reservations are terminated after host
reservations and that neutron ports are deleted if there is no
instance on the host rather than if the vif still exists in the ironic
ports internal info.