Fix invalid function signature on reallocate()
Microversion support in I7ea48be72897a77fc8424a57f4ce2d4798daf4eb added
a requirement that all API endpoints now must take the flask request as
the first argument.

Without this, we can see the following error when invoking reallocate:

  TypeError: reallocate() got multiple values for keyword argument 'host_id'

The error happens because reallocate() is called with host_id as a kwarg.
But, that argument is defined as a positional arg here. With the 'req'
arg missing, that kwarg is first interpreted as a positional arg, then
additionally as a kwarg, which confuses python.

See https://docs.python.org/3/reference/expressions.html#calls for a
description of this behavior: "If keyword arguments are present..."

Change-Id: I92f7d5fb72fa8826c801af5851d863688f61ceb0