Fix multi-region issue with network reservations (#29)
The os_region_name kwargs for the Ironic client was removed in Stein -
this means multi-region network reservations have been failing to tear
down since the Train release, as Blazar comes with an older version of
python-ironnicclient.

This updates the network_plugin to use the ironic API added as part of
the host lease healing code.

Change-Id: Ifa8684b2b2fe900f2ab35b34ddec71ea2b3d7505