Fix issue with updating extra capabilities
Updating extra capabilities was broken and generated messages like this:

    File ".../blazar/plugins/oshosts/host_plugin.py", line 525, in update_computehost
      if self.is_updatable_extra_capability(raw_capability):
    File ".../blazar/plugins/oshosts/host_plugin.py", line 491, in is_updatable_extra_capability
      capability['computehost_id'], datetime.datetime.utcnow(),
    TypeError: tuple indices must be integers, not str

The issue seems to be that the query result returned from SQLalchemy is
a list of tuples, with the first item being the plugin-specific extra
capability (e.g., `ComputeHostExtraCapability`), and the second item
being the capability name.

I initially tried removing the `add_column`, which was adding the
capability_name to the result set explicitly, but I think that's the
only way you can get this data--it's just as part of a tuple and not
part of the representation.

Change-Id: I2cdbea6a4c57ca86b2989f958ec0049375c5e28c