Fix issue fetching allocs by reservation
Filtering by lease/reservation status is incorrect in the case of
updating a lease; the lease will be in a transitioning state of
"UPDATING", which will fail the filter, and therefore it cannot find
existing reservations when constructing the request to the external
enforcement service. There doesn't seem to be any legit reason to filter
by any status here, as we're already discounting deleted reservations,
so this just removes the filter entirely. It would also fail to pick up
error leases, which are still taking up resources and should be counted
as such.

Change-Id: I16f58fc6b9f66253d0d9045e070706a273238604