Filter out unallocated hosts during alloc query
Currently, every host is returned by this query, so any lease update
request gets charged as if all resources are being reserved. The
list/query allocations stuff seems a bit brittle in terms of downstream
clients, so this is a minimal fix, just filter out the entries that
don't have any reservations; only the hosts that are currenty allocated
to the given reservation should have non-empty reservation lists.

Change-Id: I60498f00e3bea58a07eb04671ce48fc7ed354e31