Don't abort lease end event on reservation errors
If a particular reservation throws an error during its tear down, the
processing of the entire lease and all other reservations halts. This
can put the system into a bad state over time. The right thing to do is
attempt to tear down each reservation, but then raise an exception at
the end of processing all events if any of them failed.

TODO: add tests.

Change-Id: I605094fb91b549e4ab285a85e6e91b17ebe9646b