Fix empty allocations being sent to enforcement
When ending a lease, the allocations currently associated with the
lease/reservation are sent to the enforcement layers to perform
processing. This works fine when deleting a lease, but when a lease
naturally ends, that is, when the current time is > the lease's
recorded end time, the allocation lookup returns 0 results because
we have a filter on the time of the lease/reservation by default,
which starts from datetime.now. This is really not what we want
to do when explicitly filtering on a lease/reservation, so this updates
the generic query function to not apply this filtering in this case.

Change-Id: I305bb326e3bb827a927961366618b488929d5d6b