Prevent duplicate on_end enforcement calls
It turns out, it was possible for usage enforcement's on_end
hook to run multiple times if the first reservation teardown failed.
In this circumstance, the end_lease event would be IN_PROGRESS but
the reservations would be in an error state. The code only checks
if any reservation is DELETED before trying to run on_end again.

To fix this, only try to run on_end if we detected that there was
an UNDONE end_lease event; that's the only time in which we'll
actually be responsible for running this logic.

This also updates the delete logic to be a bit more consistent with
the internal way we track states. It is not enough to check dates,
because it could be that Blazar never got around to actually
starting the leases; we should be checking events. This also makes
the logic a bit easier to understand.

NOTE::
  This also fixes something which has probably been a bug for a
  very long time. The 'end' event was never set to 'DONE' when
  a lease was terminated early by the user (via a deletion.)

Change-Id: I074d75f330663dafe54ec0f75cc4c43ff56f303e