[docs] Install blazar to generate proper docs
The blazar installation was removed in [1].
However, stevedore requires blazar to generate docs for
configuration. Currently, they are mostly empty.
Instead of removing the blazar installation, this change makes
sure that deps are installed with constraints so that blazar
installation does not bring havoc and the original issue is fixed.

[1] Ic55dba0038a2209e9fb291a604a5c1da57607e1e

Change-Id: I28d06ae31c4cf44d010e0892a094138112b08641