Run unit tests in GitHub actions (#84)
* Create github actions test.yaml

* Switch to stestr

According to Openstack summit session [1],
stestr is maintained project to which all Openstack projects should migrate.
Let's switch to stestr as other projects have already moved to it.

[1] https://etherpad.openstack.org/p/YVR-python-pti

Change-Id: I9d22be39bec67fc17841367a0ef0ac439143af6c

* Fix compatibility with oslo.db 12.1.0

oslo.db 12.1.0 has changed the default value for the 'autocommit'
parameter of 'LegacyEngineFacade' from 'True' to 'False'. This is a
necessary step to ensure compatibility with SQLAlchemy 2.0. However, we
are currently relying on the autocommit behavior and need changes to
explicitly manage sessions. Until that happens, we need to override the
default.

Co-Authored-By: Stephen Finucane <stephenfin@redhat.com>
Change-Id: Ic7c5703c1d1009f6c402766aeb20fa9869495470

* Fix context unit test

This was caused by the bump of oslo.context to 4.1.0 [1], which includes
removal of the deprecated argument tenant from RequestContext [2].

Replace tenant key in the context object by project_id to fix the lower
constraints job, which is still using oslo.context 2.22.0.

[1] https://review.opendev.org/c/openstack/requirements/+/829599
[2] https://review.opendev.org/c/openstack/oslo.context/+/815938

Change-Id: I281ffa4f6fc29c691da48a792bc203818e334d9d

* run actions on push to any branch or when pull request is opened or reopened

---------

Co-authored-by: Vu Cong Tuan <tuanvc@vn.fujitsu.com>
Co-authored-by: Pierre Riteau <pierre@stackhpc.com>
Co-authored-by: Stephen Finucane <stephenfin@redhat.com>