Reduce severity of some exceptions
Many exceptions shouldn't be ERRORs in the log-file, in particular ones that don't indicate an issue in blazar's operation.

This PR is the start of a general cleanup for log-levels, with the proposed categorization:

INFO: indication of normal operation. Most enforcement exceptions should land here, as it's "normal" for a project to be low on SU, or to a lease to be rejected if too long.

WARNING: something unexpected happened, but blazar continues to operate. Lease on-end events failing could be this level.

ERROR: Something is wrong with Blazar, and an operator needs to fix it. Uncaught exceptions, un-fixable lease issues should be here. E.g. a lease-end event failed, AND retrying timed out, so the lease is stuck in an error state until an operator fixes it.

I've left all LOGs at WARN for now for consistency, but this is the overall idea.