Add validation to ensure unique capabilities when needed and device name
Added a validation method `validate_capability_value` to the `ComputeHostExtraCapability` model to
enforce uniqueness of `ExtraCapability`'s `capability_name` when its `is_unique=True`.
This validation ensures that only one capability is allowed for each resource when is_unique is set to True.
If a duplicate entry is found, a ValueError is raised with details about the non-unique capability and the affected compute host.

Set Uniqueness for `name` in `Device` model

Change-Id: I523a133abe15797cfb5b0b0edc830918dbe26f15