Add validation for capability value - updating the value
Earlier only the validation for capability value - if the value for the capability
already exists is checked when the capability is first assigned to the node.

Does not consider the case, when the value is updated for a node.
Adding another method in models to check if any node has the same value for a capability
when is_unique is True

Change-Id: I115a6de12cabc2e32b5c10669b0f9c7b0fdbb5f5