improved validation in computehost extra capability model
Previously, validation was only performed for the `capability_id` and `capability_value` attributes separately.
This update consolidates the validation logic into a single method called `validate_capability`, which is invoked for both attributes.
Additionally, the docstring for the method has been updated to provide clarity on when each validation scenario occurs.

Change-Id: I149784b5e4be71b34dfebad470cc645fed5656b4