Add new blazar api endpoint for hosts in lease
This commit addresses the complexity and inefficiency of the existing method for obtaining current nodes
in a lease in Blazar. Previously, this process involved making two large-volume API calls to
fetch all hosts and allocations in Blazar, which resulted in delays and negatively impacted downstream implementations.

To streamline this process and improve efficiency, a new API endpoint,
`reservation/v1/leases/<lease_id>/nodes`, has been introduced to Blazar leases.
This endpoint provides the Blazar host IDs and hypervisor hostnames associated with the specified lease ID.

By offering this functionality, operators and TMs can now obtain information about the nodes allocated
to a lease with a single api call, significantly reducing the time required.
Additionally, this change aims to enhance the loading speed of the lease details page in Horizon,
thereby improving the overall user experience.

Change-Id: I9271ceb1afd7057305bcccce5052a4824e9fe652