Improve host reallocation
Host reallocate could ruin a reservation if no other hosts were found,
and it could reallocate to a reserved host.

The first issue is fixed by only deallocating on success, or when the
user specifies `force`.

The second issue is fixed by never reallocating to an unreservable host,
since this behavior is never desired.

Additionally, reallocation is made easier with `reallocate_to` param,
which will try to move the allocations to this host.

Change-Id: I72e39511a89466859fa1133b409edc13a5e59800