refactor usage of placement API in flavor plugin
Instead of querying placment for all traits for all blazar hosts, instead ask
placement for a list of RPs matching the traits we want.
This should go from len(traits) * len(hosts) to 1 query.

Also include a hacky example of how to modify blazar placement client
we must query placement API with params, and current placementclient
does not implement that.